// packages/bridges/src/model.generated.ts
export const NTT_ROUTES = [
  {
    id: "wormhole-ntt-musd-mezo-to-ethereum",
    source: {
      networkId: "mezo-mainnet",
      wormholeChainId: 50,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "893d9275226a5019c7ad52628010228272d2124732d951a0ca376aded02b4107",
      decimals: 18,
      mode: 0,
    },
    destination: {
      networkId: "ethereum-mainnet",
      wormholeChainId: 2,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "ddfeb0e9ac8cf310aaa3fa4150fcca7fed33a31579df515f0041436360574441",
      decimals: 18,
      mode: 1,
    },
  },
  {
    id: "wormhole-ntt-musd-ethereum-to-mezo",
    source: {
      networkId: "ethereum-mainnet",
      wormholeChainId: 2,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "ddfeb0e9ac8cf310aaa3fa4150fcca7fed33a31579df515f0041436360574441",
      decimals: 18,
      mode: 1,
    },
    destination: {
      networkId: "mezo-mainnet",
      wormholeChainId: 50,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "893d9275226a5019c7ad52628010228272d2124732d951a0ca376aded02b4107",
      decimals: 18,
      mode: 0,
    },
  },
  {
    id: "wormhole-ntt-musd-mezo-to-base",
    source: {
      networkId: "mezo-mainnet",
      wormholeChainId: 50,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "893d9275226a5019c7ad52628010228272d2124732d951a0ca376aded02b4107",
      decimals: 18,
      mode: 0,
    },
    destination: {
      networkId: "base-mainnet",
      wormholeChainId: 30,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "2bd5ddb1e191345c82d7f824e2917f0f717316df3b86b1245680c678b2ec9417",
      decimals: 18,
      mode: 1,
    },
  },
  {
    id: "wormhole-ntt-musd-base-to-mezo",
    source: {
      networkId: "base-mainnet",
      wormholeChainId: 30,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "2bd5ddb1e191345c82d7f824e2917f0f717316df3b86b1245680c678b2ec9417",
      decimals: 18,
      mode: 1,
    },
    destination: {
      networkId: "mezo-mainnet",
      wormholeChainId: 50,
      managerId: "bridge.musd-ntt-manager",
      transceiverId: "bridge.musd-wormhole-transceiver",
      token: "0xdd468a1ddc392dcdbef6db6e34e89aa338f9f186",
      tokenCodeSha256: "893d9275226a5019c7ad52628010228272d2124732d951a0ca376aded02b4107",
      decimals: 18,
      mode: 0,
    },
  },
] as const;